pyproject.toml
[project]
name = "simple-cnn"
version = "0.1.0"
description = "A simple CNN implementation using PyTorch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "pillow",
    "numpy",
    "pandas>=2.2.3",
    "matplotlib>=3.10.1",
    "scipy==1.15",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
